Move `isomorphic-unfetch` dependency
`isomorphic-unfetch` is only used in `@appsignal/core` to perform
a push API validation request. This request is only ever performed
by the CLI package.

Move the push API validation request function to `@appsignal/cli`
alongside with the dependency on `isomorphic-unfetch`, and remove
`@appsignal/cli`'s dependency on `@appsignal/core`.
