+++ packages/cli/.changesets/remove-dependency-on---appsignal-core-.md
@@ -1,0 +1,6 @@
+---
+bump: "patch"
+type: "remove"
+---
+
+Remove dependency on `@appsignal/core`.
