+++ packages/cli/src/utils.ts
@@ -1,5 +1,6 @@
 import { existsSync } from "fs"
 import { dirname } from "path"
+import fetch from "isomorphic-unfetch"
 
 // Returns the filesystem location at which the root of the
 // `@appsignal/nodejs` package was found.
@@ -45,4 +46,25 @@ export function checkForAppsignalPackage(): string {
   console.error("Please run `npm install` and try again. Exiting.")
 
   process.exit(1)
+}
+
+export async function validatePushApiKey({
+  endpoint = "https://push.appsignal.com",
+  apiKey
+}: {
+  endpoint?: string
+  apiKey: string
+}) {
+  const { status } = await fetch(`${endpoint}/1/auth?api_key=${apiKey}`)
+
+  switch (status) {
+    case 200:
+      return true
+    case 401:
+      return false
+    default:
+      throw new Error(
+        `Invalid ${status} response from server when authenticating`
+      )
+  }
 }
