+++ packages/core/.changesets/remove-dependency-on--isomorphic-unfetch-.md
@@ -1,0 +1,7 @@
+---
+bump: "patch"
+type: "remove"
+---
+
+Remove dependency on `isomorphic-unfetch`. This fixes an issue where
+`isomorphic-unfetch` fails to bundle properly with `esbuild` or `webpack`.
