+++ packages/core/package.json
@@ -26,7 +26,6 @@
   },
   "dependencies": {
     "@appsignal/types": "=3.0.1",
-    "isomorphic-unfetch": "=4.0.0",
     "tslib": "^2.3.0"
   }
 }
