+++ packages/core/src/index.ts
@@ -5,7 +5,6 @@ export * from "./utils/error"
 export * from "./utils/url"
 export * from "./utils/async"
 export * from "./utils/environment"
-export * from "./utils/push-api"
 
 // Classes
 export { Serializable } from "./serializable"
